+++ ui/desktop/app/routes/scopes/scope/projects/sessions.js
@@ -84,6 +84,6 @@ export default class ScopesScopeProjectsSessionsRoute extends Route {
   @notifySuccess('notifications.canceled-success')
   async cancelSession(session) {
     await session.cancelSession();
-    await this.ipc.invoke('cancel', { session_id: session.id });
+    await this.ipc.invoke('stop', { session_id: session.id });
   }
 }
