+++ ui/desktop/tests/acceptance/projects/targets/sessions-test.js
@@ -271,7 +271,7 @@ module('Acceptance | projects | targets | sessions', function (hooks) {
 
   test('cancelling a session with ipc error shows notification', async function (assert) {
     assert.expect(1);
-    stubs.ipcService.withArgs('cancel').throws();
+    stubs.ipcService.withArgs('stop').throws();
     later(async () => {
       run.cancelTimers();
       await click('tbody tr:first-child td:last-child button');
